Move shared super trait code to new mod
